requirements.txt
Flask==2.2.2
google-cloud-bigquery==3.4.0
redis==4.3.4
requests==2.28.1
strawberry-graphql[debug-server]